+++ src/scss/chester/_genericCustom.scss
@@ -221,7 +221,7 @@ $enable-shadows:              false;
 $enable-gradients:            false !default;
 $enable-transitions:          true !default;
 $enable-reduced-motion:       true !default;
-$enable-smooth-scroll:        true !default;
+$enable-smooth-scroll:        false;
 $enable-grid-classes:         true !default;
 $enable-button-pointers:      true !default;
 $enable-rfs:                  true !default;
