+++ src/views/react-base.php
@@ -1,6 +1,6 @@
 <?php
 
-include BASE_PATH . "views/header.php";
+include BASE_PATH . "views/head.php";
 
 ?>
 <div id="root"></div>
@@ -8,4 +8,5 @@ include BASE_PATH . "views/header.php";
 
 $footer = new SCDS\Footer();
 $footer->addJs(getCompiledAsset('main-react-router.js'));
+$footer->showContent(false);
 $footer->render();
